+++ src/main/java/io/github/astrapi69/swing/menu/factory/MenuVisitorExtensions.java
@@ -1,0 +1,82 @@
+package io.github.astrapi69.swing.menu.factory;
+
+import java.awt.event.ActionListener;
+import java.util.Map;
+
+import javax.swing.JMenu;
+import javax.swing.JMenuBar;
+import javax.swing.JMenuItem;
+
+import io.github.astrapi69.swing.menu.enumtype.MenuType;
+import io.github.astrapi69.swing.menu.model.MenuInfo;
+import io.github.astrapi69.swing.menu.model.MenuItemInfo;
+import io.github.astrapi69.tree.BaseTreeNode;
+import lombok.NonNull;
+
+public class MenuVisitorExtensions
+{
+
+
+	public static void visitAndAddToMenu(
+		final @NonNull BaseTreeNode<MenuInfo, Long> menuInfoLongBaseTreeNode,
+		final @NonNull Map<String, JMenu> menuMap,
+		final @NonNull Map<String, JMenuItem> menuItemMap,
+		final @NonNull Map<String, JMenuBar> menuBarMap)
+	{
+		final MenuInfo menuInfo = menuInfoLongBaseTreeNode.getValue();
+		MenuType menuType = menuInfo.getType();
+		final String actionId = menuInfo.getName();
+		final BaseTreeNode<MenuInfo, Long> parent = menuInfoLongBaseTreeNode.getParent();
+		switch (menuType)
+		{
+			case MENU_ITEM :
+				final JMenuItem menuItem = menuItemMap.get(actionId);
+				if (menuMap.containsKey(parent.getValue().getName()))
+				{
+					final JMenu menu = menuMap.get(parent.getValue().getName());
+					menu.add(menuItem);
+				}
+				break;
+			case MENU :
+				final JMenu menu = menuMap.get(actionId);
+				if (menuBarMap.containsKey(parent.getValue().getName()))
+				{
+					final JMenuBar menuBar = menuBarMap.get(parent.getValue().getName());
+					menuBar.add(menu);
+				}
+				break;
+		}
+	}
+
+	public static void visitAndAddToMap(
+		final @NonNull BaseTreeNode<MenuInfo, Long> menuInfoLongBaseTreeNode,
+		final @NonNull Map<String, ActionListener> actionListenerMap,
+		final @NonNull Map<String, JMenu> menuMap,
+		final @NonNull Map<String, JMenuItem> menuItemMap,
+		final @NonNull Map<String, JMenuBar> menuBarMap)
+	{
+		final MenuInfo menuInfo = menuInfoLongBaseTreeNode.getValue();
+		if (actionListenerMap.containsKey(menuInfo.getName()))
+		{
+			final MenuItemInfo menuItemInfo = menuInfo
+				.toMenuItemInfo(actionListenerMap.get(menuInfo.getName()));
+			actionListenerMap.remove(menuInfo.getName());
+			MenuType menuType = menuInfo.getType();
+			switch (menuType)
+			{
+				case MENU_BAR :
+					final JMenuBar menuBar = menuItemInfo.toJMenuBar();
+					menuBarMap.put(menuInfo.getName(), menuBar);
+					break;
+				case MENU_ITEM :
+					final JMenuItem menuItem = menuItemInfo.toJMenuItem();
+					menuItemMap.put(menuInfo.getName(), menuItem);
+					break;
+				case MENU :
+					final JMenu menu = menuItemInfo.toJMenu();
+					menuMap.put(menuInfo.getName(), menu);
+					break;
+			}
+		}
+	}
+}